Little refactoring ExportServices
Ticket: #212
