+++ src/app/Core/ExportServices/DataExportServices/generateDataString.service.ts
@@ -1,0 +1,29 @@
+import { MainFilterLabelInterface } from "../../../Shared/Model/Export.model";
+import {
+    DbCollection,
+    DbKey,
+} from "../../../Shared/Model/Client_Isolate.model";
+import { generateDataCsvString } from "./generateDataCsvString.service";
+
+/**
+ * @desc Returns the table header and the filtered dataset as string to save it as csv
+ * @param {string[]} keys - keys of the dataset
+ * @param {Record<string, string>[]} data - dataset
+ * @returns {string} - dataset as string
+ */
+export function generateDataString(
+    data: DbCollection,
+    keys: DbKey[], 
+    mainFilterLabels: MainFilterLabelInterface,
+    mainFilterAttributes: string[],
+): string {
+    const FilteredDataString: string[] = [];
+    const headerArray: string[] = []
+    mainFilterAttributes.forEach(element => {
+        headerArray.push(mainFilterLabels[element])
+    });
+    FilteredDataString.push(headerArray.join(","));
+    FilteredDataString.push(generateDataCsvString(data, keys));
+
+    return FilteredDataString.join("\n");
+}
