+++ src/app/Core/ExportServices/generateExportLabels.service.ts
@@ -13,16 +13,16 @@ export interface ExportLabels {
  * @returns { ExportLabels } - object of labels: for the file, of the main filters, for "all values"
  */
 export function generateExportLabels(mainFilterAttributes: string[]): ExportLabels {
-    const { t } = useTranslation(["Header", "QueryPage"]);
+    const { t } = useTranslation(["QueryPage"]);
 
     const ZNFilename = `ZooNotify_${getCurrentDate()}.csv`;
 
     const mainFilterLabels = {} as MainFilterLabelInterface;
     mainFilterAttributes.forEach((mainFilter) => {
-        mainFilterLabels[mainFilter] = t(`QueryPage:Filters.${mainFilter}`);
+        mainFilterLabels[mainFilter] = t(`Filters.${mainFilter}`);
     });
 
-    const allFilterLabel: string = t("QueryPage:Filters.All");
+    const allFilterLabel: string = t("Filters.All");
 
     return { ZNFilename, mainFilterLabels, allFilterLabel };
 }
